requirements.txt
# Core dependencies - Strands with OpenAI support
strands-agents[openai]>=1.0.0
strands-agents-tools>=1.0.0
openai>=1.0.0
anthropic>=0.25.0
boto3>=1.28.0

# Graphiti and Neo4j
graphiti-core>=0.3.0
neo4j>=5.0.0
sentence-transformers>=2.2.0

# Framework and utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0

# CLI and display
typer>=0.9.0
rich>=13.0.0

# API and server
fastapi>=0.100.0
uvicorn>=0.23.0

# Token management
tiktoken>=0.5.0

# Optional: LangChain for additional capabilities
langchain>=0.1.0
langchain-community>=0.1.0

# Development tools (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0
ipython>=8.14.0